chapter 3: uh we studio'ng it
- i finally got webkit compiled and now im doing studio
- i did some refactoring of old files, and fixed the "patches" i made for some files for vs2015.
- added more linux support (UserInputService.Platform.PLATFORM_UNIX!)
- reorganised some stuff around

+++ Rendering/GfxCore/GL/ShaderGL.cpp
@@ -251,7 +251,7 @@ static UniformTable extractUniforms(unsigned int id)
 
         if (location >= 0 && uniformType != ShaderProgramGL::Uniform::Type_Unknown)
 		{
-			ShaderProgramGL::Uniform uniform = {location, uniformType, size, 0};
+			ShaderProgramGL::Uniform uniform = {location, uniformType, static_cast<unsigned int>(size), 0};
 
 			std::string name(&uniformName[0]);
